requirements.txt
absl-py==1.4.0
argcomplete==3.1.2
asgiref==3.7.2
astunparse==1.6.3
awscli==1.29.49
blinker==1.6.2
boto3==1.28.49
botocore==1.31.49
cachetools==5.3.1
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.4
contourpy==1.0.5
cycler==0.11.0
Django==4.2.4
djangorestframework==3.14.0
docutils==0.16
durationpy==0.5
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.37.2
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
hjson==3.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kappa==0.6.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.6.0
mysqlclient==2.2.0
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
placebo==0.9.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.2
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zappa==0.58.0
